#173: structure refactoring
