+++ tests/NovaTestTraitTest.php
@@ -5,9 +5,9 @@ namespace RonasIT\Support\Tests;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Route;
-use RonasIT\Support\Tests\Support\Mock\MockAuthUser;
-use RonasIT\Support\Tests\Support\Mock\TestModel;
-use RonasIT\Support\Tests\Support\Mock\TestNovaAction;
+use RonasIT\Support\Tests\Support\Mock\Actions\TestNovaAction;
+use RonasIT\Support\Tests\Support\Mock\Models\MockAuthUser;
+use RonasIT\Support\Tests\Support\Mock\Models\TestModel;
 use RonasIT\Support\Traits\MockTrait;
 use RonasIT\Support\Traits\NovaTestTrait;
 use Symfony\Component\HttpFoundation\Response;
