+++ tests/support/Mock/Actions/TestNovaAction.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace RonasIT\Support\Tests\Support\Mock\Actions;
+
+class TestNovaAction
+{
+    public function uriKey(): string
+    {
+        return 'test-nova-action';
+    }
+}
