+++ tests/support/Mock/Mails/TestMailManyFromWithName.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace RonasIT\Support\Tests\Support\Mock;
+namespace RonasIT\Support\Tests\Support\Mock\Mails;
 
 class TestMailManyFromWithName extends TestMail
 {
