+++ tests/support/Mock/Models/MockAuthUser.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace RonasIT\Support\Tests\Support\Mock;
+namespace RonasIT\Support\Tests\Support\Mock\Models;
 
 use Illuminate\Contracts\Auth\Authenticatable;
