+++ tests/support/Mock/Models/TestModel.php
@@ -1,11 +1,12 @@
 <?php
 
-namespace RonasIT\Support\Tests\Support\Mock;
+namespace RonasIT\Support\Tests\Support\Mock\Models;
 
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\HasMany;
-use RonasIT\Support\Traits\ModelTrait;
 use Illuminate\Database\Eloquent\SoftDeletes;
+use RonasIT\Support\Tests\Support\Mock\Casts\JSONCustomCast;
+use RonasIT\Support\Traits\ModelTrait;
 
 class TestModel extends Model
 {
