--- tests/support/Mock/TestNovaAction.php
@@ -1,11 +1,0 @@
-<?php
-
-namespace RonasIT\Support\Tests\Support\Mock;
-
-class TestNovaAction
-{
-    public function uriKey(): string
-    {
-        return 'test-nova-action';
-    }
-}
